Adds a basic test for the ChallengeModal POST (is the request body as expected)
