+++ e2e-tests/moderation/mod-auto-warn-first-turn-escaper.ts
@@ -22,11 +22,11 @@ import { Browser } from "@playwright/test";
 import { newTestUsername, prepareNewUser } from "@helpers/user-utils";
 
 import {
-    acceptDirectChallenge,
-    clickInTheMiddle,
     createDirectChallenge,
+    acceptDirectChallenge,
     defaultChallengeSettings,
-} from "@helpers/game-utils";
+} from "@helpers/challenge-utils";
+import { clickInTheMiddle } from "@helpers/game-utils";
 
 export const modWarnFirstTurnEscapersTest = async ({ browser }: { browser: Browser }) => {
     const { userPage: challengerPage } = await prepareNewUser(
