+++ e2e-tests/moderation/mod-block-early-stall-report.ts
@@ -21,12 +21,10 @@
 
 import { Browser, expect } from "@playwright/test";
 
-import { prepareNewUser, newTestUsername } from "@helpers/user-utils";
-import {
-    createDirectChallenge,
-    acceptDirectChallenge,
-    clickInTheMiddle,
-} from "@helpers/game-utils";
+import { newTestUsername, prepareNewUser } from "@helpers/user-utils";
+
+import { createDirectChallenge, acceptDirectChallenge } from "@helpers/challenge-utils";
+import { clickInTheMiddle } from "@helpers/game-utils";
 
 export const modBlockEarlyStallingReportTest = async ({ browser }: { browser: Browser }) => {
     const { userPage: reporterPage } = await prepareNewUser(
