+++ eslint.config.mjs
@@ -211,7 +211,6 @@ export default [
         languageOptions: {
             globals: {
                 ...globals.browser,
-                expect: "readonly",
             },
         },
     },
